+++ lib/IDatabaseConnection.d.ts
@@ -1,13 +1,14 @@
 /// <reference types="node" />
 import { Readable } from 'stream';
+import { Query } from './Query';
 export interface IDatabaseConnection {
     getInstantiationStack(): string;
     getAPI(): any;
     isReadOnly(): boolean;
     setTimeout(timeout: number): void;
     getTimeout(): number;
-    query(query: string, params?: any): Promise<any>;
-    stream(query: string, params?: any, streamOptions?: any): Readable;
+    query(query: string | Query, params?: any): Promise<any>;
+    stream(query: string | Query, params?: any, streamOptions?: any): Readable;
     close(forceClose?: boolean): Promise<void>;
     startTransaction(): Promise<void>;
     isTransaction(): boolean;