+++ src/IDatabaseConnection.ts
@@ -14,6 +14,7 @@
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 import { Readable } from 'stream';
+import { Query } from './Query';
 
 export interface IDatabaseConnection {
     getInstantiationStack(): string;
@@ -21,8 +22,8 @@ export interface IDatabaseConnection {
     isReadOnly(): boolean;
     setTimeout(timeout: number): void;
     getTimeout(): number;
-    query(query: string, params?: any): Promise<any>;
-    stream(query: string, params?: any, streamOptions?: any): Readable;
+    query(query: string | Query, params?: any): Promise<any>;
+    stream(query: string | Query, params?: any, streamOptions?: any): Readable;
     close(forceClose?: boolean): Promise<void>;
     startTransaction(): Promise<void>;
     isTransaction(): boolean;